Migrate water/data tests to appropriate location
