+++ message_ix_models/tests/model/water/data/test_water_supply.py
